--- a/README.md
+++ b/README.md
@@ -2,12 +2,11 @@ Two ringbuffers written in Go.
 
 ### ringbuffer1
 
-[ringbuffer1](./ringbuffer1) stores bytes. It implements an indexing strategy learned from https://www.snellman.net/blog/archive/2016-12-13-ring-buffers/
+[ringbuffer1](./ringbuffer1) stores bytes. It features:
 
-Dependencies for ringbuffer1 are for running tests:
-
-* https://github.com/flyingmutant/rapid
-* https://github.com/google/gofuzz
+1. indexing strategy learned from https://www.snellman.net/blog/archive/2016-12-13-ring-buffers/
+2. https://github.com/bmkessler/fastdiv for faster modulo on the read/write indices
+2. https://github.com/google/gofuzz and https://github.com/flyingmutant/rapid for testing
 
 ### ringbuffer2