--- a/ringbuffer1/go.mod
+++ b/ringbuffer1/go.mod
@@ -3,6 +3,7 @@ module github.com/sevagh/ringworm/ringbuffer1
 go 1.13
 
 require (
+	github.com/bmkessler/fastdiv v0.0.0-20190227075523-41d5178f2044
 	github.com/flyingmutant/rapid v0.0.0-20190904072629-5761511f78c8
 	github.com/google/gofuzz v1.0.0
 )