By using the fastdiv library (https://github.com/bmkessler/fastdiv, based on https://github.com/lemire/fastmod), I can support arbitrary ringbuffer sizes without the expensive modulo (currently I enforce power-of-two to use the `a & (b - 1)` trick).

Benchmarks:
```
sevagh:ringbuffer1 $ benchcmp old.txt fastdiv.txt
benchmark                                        old ns/op       new ns/op       delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     41697812531     47722719778     +14.45%

benchmark                                        old allocs     new allocs     delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     2              4              +100.00%

benchmark                                        old bytes     new bytes     delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     96            520           +441.67%
sevagh:ringbuffer1 $
sevagh:ringbuffer1 $ benchcmp old.txt regular_modulo.txt
benchmark                                        old ns/op       new ns/op       delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     41697812531     66041473288     +58.38%

benchmark                                        old allocs     new allocs     delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     2              2              +0.00%

benchmark                                        old bytes     new bytes     delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     96            96            +0.00%
sevagh:ringbuffer1 $
sevagh:ringbuffer1 $ benchcmp regular_modulo.txt fastdiv.txt
benchmark                                        old ns/op       new ns/op       delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     66041473288     47722719778     -27.74%

benchmark                                        old allocs     new allocs     delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     2              4              +100.00%

benchmark                                        old bytes     new bytes     delta
BenchmarkManyRingbuffersBillionsOfIntegers-8     96            520           +441.67%
```